+++ src/Support/RuleListTrait.php
@@ -10,9 +10,9 @@ trait RuleListTrait
      * @var array
      */
     protected $clientRules = ['Accepted', 'After', 'Alpha', 'AlphaDash',
-        'AlphaNum', 'Array', 'Bail', 'Before', 'Between', 'Boolean', 'Confirmed', 'Date',
-        'DateFormat', 'Different', 'Digits', 'DigitsBetween', 'Email', 'Image',
-        'In', 'Integer', 'Ip', 'Json', 'Max', 'Mimes', 'Min', 'NotIn', 'Numeric',
+        'AlphaNum', 'Array', 'Bail', 'Before', 'Between', 'Boolean', 'Confirmed', 'Date', 'Dimensions',
+        'DateFormat', 'Different', 'Digits', 'DigitsBetween', 'Email', 'File', 'Filled','Image',
+        'In', 'Integer', 'Ip', 'Json', 'Max', 'Mimes', 'Mimetypes', 'Min', 'NotIn', 'Nullable', 'Numeric',
         'Regex', 'Required', 'RequiredIf', 'RequiredUnless', 'RequiredWith', 'RequiredWithAll',
         'RequiredWithout', 'RequiredWithoutAll', 'Same', 'Size', 'Sometimes',
         'String', 'Timezone', 'Url', ];
@@ -29,7 +29,7 @@ trait RuleListTrait
      *
      * @var array
      */
-    protected $fileRules = ['Image', 'Mimes'];
+    protected $fileRules = ['File', 'Image', 'Mimes', 'Mimetypes'];
 
     /**
      * Rule used to disable validations.