+++ tests/Remote/ValidatorTest.php
@@ -36,6 +36,9 @@ class ValidatorTest extends \PHPUnit_Framework_TestCase
         try {
             $validator->validate('_jsvalidation','field',[]);
             $this->fail();
+        } catch (\Illuminate\Validation\ValidationException $ex) {
+            $this->assertEquals(200, $ex->getResponse()->getStatusCode());
+            $this->assertEquals('["field active_url!"]', $ex->getResponse()->getContent());
         } catch (HttpResponseException $ex) {
             $this->assertEquals(200, $ex->getResponse()->getStatusCode());
             $this->assertEquals('["field active_url!"]', $ex->getResponse()->getContent());
@@ -66,6 +69,9 @@ class ValidatorTest extends \PHPUnit_Framework_TestCase
             $validator->setValidateAll(true);
             $validator->validate('_jsvalidation','field',[]);
             $this->fail();
+        } catch (\Illuminate\Validation\ValidationException $ex) {
+            $this->assertEquals(200, $ex->getResponse()->getStatusCode());
+            $this->assertEquals('["validation.alpha"]', $ex->getResponse()->getContent());
         } catch (HttpResponseException $ex) {
             $this->assertEquals(200, $ex->getResponse()->getStatusCode());
             $this->assertEquals('["validation.alpha"]', $ex->getResponse()->getContent());